load spct to table

+++ src/main/java/org/project1/nhom8/model/GiaModel.java
@@ -38,7 +38,7 @@ public class GiaModel {
     private int maSPCT;
 
     @DataField(name = "GIA")
-    private float gia;
+    private double gia;
 
     @DataField(name = "NGAYUPDATE")
     private Date ngayCapNhat;
